Added new OV parallel
